backend/requirements.txt
Django
psycopg[binary]
python-dotenv
django-cors-headers
django-debug-toolbar
graphene-django
django-graphql-jwt
coverage
